fix: correção de erros na tela de perfil
